requirements.txt
Django==3.2.12
psycopg2==2.9.3
django-cors-headers>=3.11.0
requests>=2.27.1
django-filter==22.1
django-request-logging>=0.7.5
python-json-logger>=0.1.11
firebase-admin>=5.2.0
python-dateutil==2.8.2
pillow>=9.2.0
mysqlclient>=2.1.1
graphene-django==2.15.0
graphene~=2.1.9
django-graphql-jwt==0.3.4
django-admin-list-filter-dropdown==1.0.3
graphene-file-upload==1.3.0
djangorestframework==3.13.1
django-extensions==3.2.1
django-apscheduler == 0.6.2
pandas==1.5.1
openpyxl==3.0.10
celery>=5.2
Redis>=4.3.4

